Abins 2D: Break out some common 1D/2D features into a new file.
Use multiple inheritance(!) to set up some private methods that
provide the common stuff and keep 1D and 2D algorithms fairly lean
and easy to a) compare b) maintain
